Removed dead l2chain code. Moved chain_test to use block consmer component albeit the test is sketchy.
