+++ go/enclave/components/consumer_test.go
@@ -1,28 +1,19 @@
-package l2chain
+package components
 
 import (
 	"math/big"
 	"testing"
 
-	"github.com/obscuronet/go-obscuro/go/common/log"
-
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/trie"
-
-	gethlog "github.com/ethereum/go-ethereum/log"
 )
 
 func TestInvalidBlocksAreRejected(t *testing.T) {
-	// There are no tests of acceptance of valid chains of blocks. This is because the logic to generate a valid block
-	// is non-trivial.
-	genesisJSON, err := core.DefaultGenesisBlock().MarshalJSON()
-	if err != nil {
-		t.Errorf("could not parse genesis JSON: %v", err)
-	}
-	logger := log.New(log.DeployerCmp, int(gethlog.LvlDebug), log.SysOut)
-	chain := ObscuroChain{l1Blockchain: NewL1Blockchain(genesisJSON, logger)}
+
+	//todo - how does this test even work, storage is never set and we attempt to fetch head block?
+	blockConsumer := blockConsumer{}
 
 	invalidHeaders := []types.Header{
 		{ParentHash: common.HexToHash("0x0")},                                                            // Unknown ancestor.
@@ -32,7 +23,7 @@ func TestInvalidBlocksAreRejected(t *testing.T) {
 
 	for _, header := range invalidHeaders {
 		loopHeader := header
-		_, err := chain.insertBlockIntoL1Chain(types.NewBlock(&loopHeader, nil, nil, nil, &trie.StackTrie{}), false)
+		_, err := blockConsumer.ingestBlock(types.NewBlock(&loopHeader, nil, nil, nil, &trie.StackTrie{}), false)
 		if err == nil {
 			t.Errorf("expected block with invalid header to be rejected but was accepted")
 		}
